+++ Runtime/IStateProviderSource.cs
@@ -1,7 +1,0 @@
-namespace UniMob.UI
-{
-    public interface IStateProviderSource
-    {
-        IStateProvider StateProvider { get; }
-    }
-}