+++ Runtime/UniMobUIApp.cs
@@ -8,13 +8,11 @@ namespace UniMob.UI
         [SerializeField]
         private ViewPanel root = null!;
 
-        public StateProvider StateProvider { get; } = new StateProvider();
-
         protected override void Start()
         {
             Initialize();
 
-            UniMobUI.RunApp(Lifetime, StateProvider, root, Build, name);
+            UniMobUI.RunApp(Lifetime, root, Build, name);
         }
 
         protected virtual void Initialize() { }