+++ Testing/CountingBox.cs
@@ -22,8 +22,6 @@ namespace UniMob.UI.Tests
         public System.Type Type => typeof(CountingBox);
         public Key Key { get; set; }
 
-        public State CreateState(StateProvider provider) => null;
-
         public State CreateState() => new CountingBoxState();
 
         public RenderObject CreateRenderObject(BuildContext context, IState state) =>