+++ Testing/WidthProbeBox.cs
@@ -16,8 +16,6 @@ namespace UniMob.UI.Tests
         public System.Type Type => typeof(WidthProbeBox);
         public Key Key { get; set; }
 
-        public State CreateState(StateProvider provider) => null;
-
         public State CreateState() => new WidthProbeBoxState();
 
         public RenderObject CreateRenderObject(BuildContext context, IState state) =>